do not attempt to delete undefined head from storage

+++ packages/datastore/src/sync/outbox.ts
@@ -80,8 +80,9 @@ class MutationEventOutbox {
 	public async dequeue(storage: StorageFacade): Promise<MutationEvent> {
 		const head = await this.peek(storage);
 
-		await storage.delete(head);
-
+		if (head !== undefined) {
+			await storage.delete(head);
+		}
 		this.inProgressMutationEventId = undefined;
 
 		return head;
